Remove unreachable remote refs (#16)
The $ref values I'm removing are part of components.examples that I think are only used in the responses. Examples is optional and its structure is kind of open form.

These removed refs point to external URLs that are unreachable or contain a response that a library we use doesn't know how to interpret. In doc-gen, we are using a parser library that tries to get the values from these remote URLs and it fails to do so.

An example of one of the removed refs:
"$ref": "https://api.github.com/scim/v2/enterprises/octo-corp/Users/92b58aaa-a1d6-11ea-8227-b9ce9e023ccc",

Kept the original GH specs intact and instead modified the new copies.