Rename fixed specs to be more explicit (#17)
The fix done on these specs (PR #16) is removing some invalid $ref values. Rename the files accordingly so its easier to remember what the fix was.

These files will be removed eventually when doc-gen is ready to let you know what the errors are in your spec or why it cannot parse a particular spec.